+++ src/date_utils.lua
@@ -1,7 +1,15 @@
 local date_utils = {}
 
 function date_utils.get_hours_and_minutes(minutes)
-   return math.floor(minutes / 60), math.floor((minutes % 60) + 0.5)
+   local hours = math.floor(minutes / 60)
+   local mins = math.floor((minutes % 60) + 0.5)
+
+   if minutes == 60 then
+      hours = hours + 1
+      mins = 0
+   end
+
+   return hours, mins
 end
 
 function date_utils.date_string_to_timestamp(date_string)